Crate rusoto_acm

AWS Certificate Manager
Welcome to the AWS Certificate Manager (ACM) API documentation.

You can use ACM to manage SSL/TLS certificates for your AWS-based websites and applications. For general information about using ACM, see the AWS Certificate Manager User Guide .

If you're using the service, you're probably looking for AcmClient and Acm.

Structs

AcmClient
A client for the ACM API.

AddTagsToCertificateRequest

CertificateDetail
Contains metadata about an ACM certificate. This structure is returned in the response to a DescribeCertificate request.

CertificateSummary
This structure is returned in the response object of ListCertificates action.

DeleteCertificateRequest

DescribeCertificateRequest

DescribeCertificateResponse

DomainValidation
Contains information about the validation of each domain name in the certificate.

DomainValidationOption
Contains information about the domain names that you want ACM to use to send you emails to validate your ownership of the domain.

GetCertificateRequest

GetCertificateResponse

ImportCertificateRequest

ImportCertificateResponse

ListCertificatesRequest

ListCertificatesResponse

ListTagsForCertificateRequest

ListTagsForCertificateResponse

RemoveTagsFromCertificateRequest

RenewalSummary
Contains information about the status of ACM's managed renewal for the certificate. This structure exists only when the certificate type is AMAZON_ISSUED.

RequestCertificateRequest

RequestCertificateResponse

ResendValidationEmailRequest

Tag
A key-value pair that identifies or specifies metadata about an ACM resource.

Enums

AddTagsToCertificateError
Errors returned by AddTagsToCertificate

DeleteCertificateError
Errors returned by DeleteCertificate

DescribeCertificateError
Errors returned by DescribeCertificate

GetCertificateError
Errors returned by GetCertificate

ImportCertificateError
Errors returned by ImportCertificate

ListCertificatesError
Errors returned by ListCertificates

ListTagsForCertificateError
Errors returned by ListTagsForCertificate

RemoveTagsFromCertificateError
Errors returned by RemoveTagsFromCertificate

RequestCertificateError
Errors returned by RequestCertificate

ResendValidationEmailError
Errors returned by ResendValidationEmail

Traits

Acm
Trait representing the capabilities of the ACM API. ACM clients implement this trait.