Crate rusoto_acm

AWS Certificate Manager
Welcome to the AWS Certificate Manager (ACM) API documentation.

You can use ACM to manage SSL/TLS certificates for your AWS-based websites and applications. For general information about using ACM, see the AWS Certificate Manager User Guide .

If you're using the service, you're probably looking for AcmClient and Acm.

Structs

AcmClient
A client for the ACM API.

AddTagsToCertificateRequest

CertificateDetail
Contains metadata about an ACM certificate. This structure is returned in the response to a DescribeCertificate request.

CertificateOptions
Structure that contains options for your certificate. Currently, you can use this only to specify whether to opt in to or out of certificate transparency logging. Some browsers require that public certificates issued for your domain be recorded in a log. Certificates that are not logged typically generate a browser error. Transparency makes it possible for you to detect SSL/TLS certificates that have been mistakenly or maliciously issued for your domain. For general information, see Certificate Transparency Logging.

CertificateSummary
This structure is returned in the response object of ListCertificates action.

DeleteCertificateRequest

DescribeCertificateRequest

DescribeCertificateResponse

DomainValidation
Contains information about the validation of each domain name in the certificate.

DomainValidationOption
Contains information about the domain names that you want ACM to use to send you emails that enable you to validate domain ownership.

ExportCertificateRequest

ExportCertificateResponse

ExtendedKeyUsage
The Extended Key Usage X.509 v3 extension defines one or more purposes for which the public key can be used. This is in addition to or in place of the basic purposes specified by the Key Usage extension.

Filters
This structure can be used in the ListCertificates action to filter the output of the certificate list.

GetCertificateRequest

GetCertificateResponse

ImportCertificateRequest

ImportCertificateResponse

KeyUsage
The Key Usage X.509 v3 extension defines the purpose of the public key contained in the certificate.

ListCertificatesRequest

ListCertificatesResponse

ListTagsForCertificateRequest

ListTagsForCertificateResponse

RemoveTagsFromCertificateRequest

RenewalSummary
Contains information about the status of ACM's managed renewal for the certificate. This structure exists only when the certificate type is AMAZON_ISSUED.

RequestCertificateRequest

RequestCertificateResponse

ResendValidationEmailRequest

ResourceRecord
Contains a DNS record value that you can use to can use to validate ownership or control of a domain. This is used by the DescribeCertificate action.

Tag
A key-value pair that identifies or specifies metadata about an ACM resource.

UpdateCertificateOptionsRequest

Enums

AddTagsToCertificateError
Errors returned by AddTagsToCertificate

DeleteCertificateError
Errors returned by DeleteCertificate

DescribeCertificateError
Errors returned by DescribeCertificate

ExportCertificateError
Errors returned by ExportCertificate

GetCertificateError
Errors returned by GetCertificate

ImportCertificateError
Errors returned by ImportCertificate

ListCertificatesError
Errors returned by ListCertificates

ListTagsForCertificateError
Errors returned by ListTagsForCertificate

RemoveTagsFromCertificateError
Errors returned by RemoveTagsFromCertificate

RequestCertificateError
Errors returned by RequestCertificate

ResendValidationEmailError
Errors returned by ResendValidationEmail

UpdateCertificateOptionsError
Errors returned by UpdateCertificateOptions

Traits

Acm
Trait representing the capabilities of the ACM API. ACM clients implement this trait.