Struct rusoto_acm::AcmClient

pub struct AcmClient { /* fields omitted */ }

A client for the ACM API.

Methods

impl AcmClient

pub fn new(region: Region) -> AcmClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> AcmClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,