Struct rusoto_acm::CertificateDetail

pub struct CertificateDetail {
    pub certificate_arn: Option<String>,
    pub certificate_authority_arn: Option<String>,
    pub created_at: Option<f64>,
    pub domain_name: Option<String>,
    pub domain_validation_options: Option<Vec<DomainValidation>>,
    pub extended_key_usages: Option<Vec<ExtendedKeyUsage>>,
    pub failure_reason: Option<String>,
    pub imported_at: Option<f64>,
    pub in_use_by: Option<Vec<String>>,
    pub issued_at: Option<f64>,
    pub issuer: Option<String>,
    pub key_algorithm: Option<String>,
    pub key_usages: Option<Vec<KeyUsage>>,
    pub not_after: Option<f64>,
    pub not_before: Option<f64>,
    pub options: Option<CertificateOptions>,
    pub renewal_eligibility: Option<String>,
    pub renewal_summary: Option<RenewalSummary>,
    pub revocation_reason: Option<String>,
    pub revoked_at: Option<f64>,
    pub serial: Option<String>,
    pub signature_algorithm: Option<String>,
    pub status: Option<String>,
    pub subject: Option<String>,
    pub subject_alternative_names: Option<Vec<String>>,
    pub type_: Option<String>,
}

Contains metadata about an ACM certificate. This structure is returned in the response to a DescribeCertificate request.

Fields
certificate_arn: Option<String>

The Amazon Resource Name (ARN) of the certificate. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

certificate_authority_arn: Option<String>

The Amazon Resource Name (ARN) of the ACM PCA private certificate authority (CA) that issued the certificate. This has the following format:

arn:aws:acm-pca:region:account:certificate-authority/12345678-1234-1234-1234-123456789012

created_at: Option<f64>

The time at which the certificate was requested. This value exists only when the certificate type is AMAZON_ISSUED.

domain_name: Option<String>

The fully qualified domain name for the certificate, such as www.example.com or example.com.

domain_validation_options: Option<Vec<DomainValidation>>

Contains information about the initial validation of each domain name that occurs as a result of the RequestCertificate request. This field exists only when the certificate type is AMAZON_ISSUED.

extended_key_usages: Option<Vec<ExtendedKeyUsage>>

Contains a list of Extended Key Usage X.509 v3 extension objects. Each object specifies a purpose for which the certificate public key can be used and consists of a name and an object identifier (OID).

failure_reason: Option<String>

The reason the certificate request failed. This value exists only when the certificate status is FAILED. For more information, see Certificate Request Failed in the AWS Certificate Manager User Guide.

imported_at: Option<f64>

The date and time at which the certificate was imported. This value exists only when the certificate type is IMPORTED.

in_use_by: Option<Vec<String>>

A list of ARNs for the AWS resources that are using the certificate. A certificate can be used by multiple AWS resources.

issued_at: Option<f64>

The time at which the certificate was issued. This value exists only when the certificate type is AMAZON_ISSUED.

issuer: Option<String>

The name of the certificate authority that issued and signed the certificate.

key_algorithm: Option<String>

The algorithm that was used to generate the public-private key pair.

key_usages: Option<Vec<KeyUsage>>

A list of Key Usage X.509 v3 extension objects. Each object is a string value that identifies the purpose of the public key contained in the certificate. Possible extension values include DIGITAL_SIGNATURE, KEY_ENCHIPHERMENT, NON_REPUDIATION, and more.

not_after: Option<f64>

The time after which the certificate is not valid.

not_before: Option<f64>

The time before which the certificate is not valid.

options: Option<CertificateOptions>

Value that specifies whether to add the certificate to a transparency log. Certificate transparency makes it possible to detect SSL certificates that have been mistakenly or maliciously issued. A browser might respond to certificate that has not been logged by showing an error message. The logs are cryptographically secure.

renewal_eligibility: Option<String>

Specifies whether the certificate is eligible for renewal. At this time, only exported private certificates can be renewed with the RenewCertificate command.

renewal_summary: Option<RenewalSummary>

Contains information about the status of ACM's managed renewal for the certificate. This field exists only when the certificate type is AMAZON_ISSUED.

revocation_reason: Option<String>

The reason the certificate was revoked. This value exists only when the certificate status is REVOKED.

revoked_at: Option<f64>

The time at which the certificate was revoked. This value exists only when the certificate status is REVOKED.

serial: Option<String>

The serial number of the certificate.

signature_algorithm: Option<String>

The algorithm that was used to sign the certificate.

status: Option<String>

The status of the certificate.

subject: Option<String>

The name of the entity that is associated with the public key contained in the certificate.

subject_alternative_names: Option<Vec<String>>

One or more domain names (subject alternative names) included in the certificate. This list contains the domain names that are bound to the public key that is contained in the certificate. The subject alternative names include the canonical domain name (CN) of the certificate and additional domain names that can be used to connect to the website.

type_: Option<String>

The source of the certificate. For certificates provided by ACM, this value is AMAZON_ISSUED. For certificates that you imported with ImportCertificate, this value is IMPORTED. ACM does not provide managed renewal for imported certificates. For more information about the differences between certificates that you import and those that ACM provides, see Importing Certificates in the AWS Certificate Manager User Guide.